feat: public recipe add and remove functionality

+++ src/composables/useDataService.ts
@@ -7,10 +7,11 @@ import {
   setDoc
 } from 'firebase/firestore'
 import type { LocalUser } from '@/types/UserState'
-import dummyData from '../assets/dummyData.json'
 import { publicRecipesRef, usersRef } from '../../firebase'
 import { Recipe } from '@/types/Recipes'
 import { useRecipeStore } from '@/stores/recipe'
+import { useUserStore } from '@/stores/user'
+import { getAuth } from 'firebase/auth'
 
 export function useDataService() {
 
@@ -31,27 +32,36 @@ export function useDataService() {
     }
   }
 
-  const alterPublicRecipesData = async () => {
+  const updatePublicRecipesData = async () => {
     const newPublicRecipes = recipeStore.newPublicRecipes
     const toBeDeletedPublicRecipes = recipeStore.removedPublicRecipes
-    if(newPublicRecipes) savePublicRecipesData(newPublicRecipes)
-    if (toBeDeletedPublicRecipes) deletePublicRecipesData(toBeDeletedPublicRecipes)
+    if(newPublicRecipes.length > 0) savePublicRecipesData(newPublicRecipes)
+    if (toBeDeletedPublicRecipes.length > 0) deletePublicRecipesData(toBeDeletedPublicRecipes)
   }
 
   const savePublicRecipesData = async (publicRecipes: Recipe[]) => {
     error.value = null // Reset error
+    const userStore = useUserStore()
     try {
       console.log('Saving public recipe data: ', publicRecipes)
-  
       for (const recipe of publicRecipes) {
+        const prefixedRecipeId = `pub-${recipe.id}`
+        const userId = userStore.getCurrentUserId
+        const updatedRecipe =  {...recipe, id: prefixedRecipeId, creatorId: userId}
+        console.log('updatedRecipe: ', updatedRecipe)
+        console.log('current User Id: ', userStore.getCurrentUserId)
         // Check if recipe id exists
-        const existingRecipeDocRef = doc(publicRecipesRef, `pub-${recipe.id}`);
-        const existingRecipe = await getDoc(existingRecipeDocRef);
-        const publicRecipeId = !existingRecipe ? `pub-${recipe.id}` : `pub-${recipe.id}-z`
-        const updatedRecipe = {...recipe, id: publicRecipeId}
-
-        await setDoc(existingRecipeDocRef, updatedRecipe);
+        const existingRecipeDocRef = doc(publicRecipesRef, prefixedRecipeId);
+        const docSnap = await getDoc(existingRecipeDocRef);
+          console.log('check for existing recipe')
+        if (docSnap.exists()) {
+          // TODO if recipe exists, display error
+          console.log('Public Recipe Already Exists')
+        } else {
+          await setDoc(existingRecipeDocRef, updatedRecipe);
+        }
       }
+      recipeStore.resetNewPublicRecipes()
     } catch (err: any) {
       error.value = err
       console.error('Error saving public Recipe data:', error.value)
@@ -61,16 +71,24 @@ export function useDataService() {
   const deletePublicRecipesData = async (publicRecipes: Recipe[]) => {
     error.value = null 
     try {
-      for (const recipe of publicRecipes) {    
-        const recipeRef = doc(publicRecipesRef, recipe.id)
-        
-        await deleteDoc(recipeRef)
+      for (const recipe of publicRecipes) {   
+        // check it existss
+        console.log('auth: ', getAuth())
+        const recipeRef = doc(publicRecipesRef, `pub-${recipe.id}`)
+        const recipeSnapshot = await getDoc(recipeRef);
+
+        if (recipeSnapshot.exists()) {
+          console.log('recipe exists in db')
+          await deleteDoc(recipeRef)
+        } else {
+          console.log('Recipes not there')
+        }
   
       }
-
+      recipeStore.resetRemovedPublicRecipes()
     } catch (err: any) {
       error.value = err
-      console.error('Error saving public Recipe data:', error.value)
+      console.error('Error deleting public Recipe data:', error.value)
     }
     
   }
@@ -114,5 +132,5 @@ const loadPublicRecipeData = async (uid: string): Promise<[DocumentData | null]>
   }
 };
 
-  return { saveUserData, loadUserData, loadPublicRecipeData, savePublicRecipesData, deletePublicRecipesData, alterPublicRecipesData, error }
+  return { saveUserData, loadUserData, loadPublicRecipeData, savePublicRecipesData, deletePublicRecipesData, updatePublicRecipesData, error }
 }
